Refactor config and fill dataclasses into dedicated modules
