+++ simulation/__init__.py
@@ -1,6 +1,7 @@
 """Simulation package exports."""
 
-from simulation.order_processor import Fill, OrderProcessor
+from simulation.models.fill import Fill
+from simulation.order_processor import OrderProcessor
 from simulation.position_simulator import StatefulPositionSimulator
 from simulation.trade_classifier import TradeClassifier
