+++ simulation/models/__init__.py
@@ -1,0 +1,5 @@
+"""Simulation models exports."""
+
+from simulation.models.fill import Fill
+
+__all__ = ["Fill"]
