+++ simulation/models/fill.py
@@ -1,0 +1,13 @@
+"""Order fill model."""
+
+from __future__ import annotations
+
+from dataclasses import dataclass
+
+
+@dataclass(frozen=True, slots=True)
+class Fill:
+    """Single order fill snapshot."""
+
+    price: float
+    commission: float
